Add `RichMessage` subtype for `Message` and `PaidMedia` subtype for `ExternalReplyInfo`. Improve integer type handling by description.

+++ telers/src/types/business_opening_hours_interval.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct BusinessOpeningHoursInterval {
     /// The minute's sequence number in a week, starting on Monday, marking the start of the time interval during which the business is open; 0 - 7 * 24 * 60
-    pub opening_minute: u8,
+    pub opening_minute: i64,
     /// The minute's sequence number in a week, starting on Monday, marking the end of the time interval during which the business is open; 0 - 8 * 24 * 60
-    pub closing_minute: u8,
+    pub closing_minute: i64,
 }
 impl BusinessOpeningHoursInterval {
     /// Creates a new `BusinessOpeningHoursInterval`.
@@ -16,7 +16,7 @@ impl BusinessOpeningHoursInterval {
     /// * `opening_minute` - The minute's sequence number in a week, starting on Monday, marking the start of the time interval during which the business is open; 0 - 7 * 24 * 60
     /// * `closing_minute` - The minute's sequence number in a week, starting on Monday, marking the end of the time interval during which the business is open; 0 - 8 * 24 * 60
     #[must_use]
-    pub fn new<T0: Into<u8>, T1: Into<u8>>(opening_minute: T0, closing_minute: T1) -> Self {
+    pub fn new<T0: Into<i64>, T1: Into<i64>>(opening_minute: T0, closing_minute: T1) -> Self {
         Self {
             opening_minute: opening_minute.into(),
             closing_minute: closing_minute.into(),
@@ -25,14 +25,14 @@ impl BusinessOpeningHoursInterval {
 
     /// The minute's sequence number in a week, starting on Monday, marking the start of the time interval during which the business is open; 0 - 7 * 24 * 60
     #[must_use]
-    pub fn opening_minute<T: Into<u8>>(mut self, val: T) -> Self {
+    pub fn opening_minute<T: Into<i64>>(mut self, val: T) -> Self {
         self.opening_minute = val.into();
         self
     }
 
     /// The minute's sequence number in a week, starting on Monday, marking the end of the time interval during which the business is open; 0 - 8 * 24 * 60
     #[must_use]
-    pub fn closing_minute<T: Into<u8>>(mut self, val: T) -> Self {
+    pub fn closing_minute<T: Into<i64>>(mut self, val: T) -> Self {
         self.closing_minute = val.into();
         self
     }
